Rename structs to lm_structs
