+++ src/lm_structs/roomcache.rs
